// fixtures/client-adapters/src/generated/api.generated.ts
export interface Widget {
  readonly id: string;
  readonly name: string;
}
// ...
export type ApiClient = ReturnType<typeof createApiClient>;